# frontend/api_doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Movie Review System API",
    "description": "Backend API for Movie Review Website",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/rating/update": {
      "post": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Update Rating",
        "operationId": "updateRating",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/rating/submit": {
      "post": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Submit Rating",
        "operationId": "submitRating",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/rating/my": {
      "get": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Get My Ratings",
        "operationId": "getMyRatings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageInfoMyRatingVO"
                }
              }
            }
          }
        }
      }
    },
    "/rating/get": {
      "get": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Get User Rating",
        "operationId": "getUserRating",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRating"
                }
              }
            }
          }
        }
      }
    },
...
    "/rating/clear": {
      "delete": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Clear All My Ratings",
        "operationId": "clearMyRatings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/rating/batch": {
      "delete": {
        "tags": [
          "Rating Management"
        ],
        "summary": "Batch Delete Ratings",
        "operationId": "deleteRatingsBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movieIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID (唯一账号)",
            "example": "movie_fan_01"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像 URL",
            "example": "http://localhost:8080/images/default_avatar.png"
          },
          "url": {
            "type": "string",
            "description": "个人主页/豆瓣链接",
            "example": "https://www.douban.com/people/xxxx"
          },
          "email": {
            "type": "string",
            "description": "电子邮箱",
            "example": "user@example.com"
          },
          "role": {
            "type": "integer",
            "description": "用户角色 (0:管理员, 1:普通用户)",
            "format": "int32",
            "example": 1
          },
          "status": {
            "type": "integer",
            "description": "账号状态 (0:正常, 1:被禁用/冻结)",
            "format": "int32",
            "example": 0
          },
          "passwordVersion": {
            "type": "integer",
            "description": "密码版本号，用于失效旧 Token（修改密码后递增）",
            "format": "int32",
            "example": 1
          },
          "createTime": {
            "type": "string",
            "description": "账号创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          }
        },
        "description": "用户实体"
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "string",
            "description": "业务数据载荷"
          }
        },
        "description": "统一 API 响应结构"
      },
...
      "MyRatingVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "评分记录ID (格式: userId_movieId)",
            "example": "user1_123456"
          },
          "movieId": {
            "type": "integer",
            "description": "电影ID",
            "format": "int64",
            "example": 1292052
          },
          "movieName": {
            "type": "string",
            "description": "电影名称",
            "example": "肖申克的救赎"
          },
          "posterUrl": {
            "type": "string",
            "description": "电影海报URL",
            "example": "http://localhost:8080/images/poster_01.jpg"
          },
          "rating": {
            "type": "integer",
            "description": "我的评分 (1-5)",
            "format": "int32",
            "example": 5
          },
          "ratingTime": {
            "type": "string",
            "description": "评分时间",
            "example": "2023-10-01 12:00:00"
          },
          "description": {
            "type": "string",
            "description": "电影简介",
            "example": "20世纪40年代末，小有成就的青年银行家安迪..."
          }
        },
        "description": "我的评分记录 (包含电影信息)"
      },
      "PageInfoMyRatingVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyRatingVO"
            }
          },
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "startRow": {
            "type": "integer",
            "format": "int64"
          },
          "endRow": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "prePage": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "integer",
            "format": "int32"
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "navigatePages": {
            "type": "integer",
            "format": "int32"
          },
          "navigatepageNums": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "navigateFirstPage": {
            "type": "integer",
            "format": "int32"
          },
          "navigateLastPage": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "业务数据载荷"
      },
      "ResultPageInfoMyRatingVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PageInfoMyRatingVO"
          }
        },
        "description": "统一 API 响应结构"
      },
      "Rating": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "movie_fan_01"
          },
          "movieId": {
            "type": "integer",
            "description": "电影ID",
            "format": "int64",
            "example": 1292052
          },
          "rating": {
            "type": "integer",
            "description": "评分值 (1-5 星)",
            "format": "int32",
            "example": 5
          },
          "ratingTime": {
            "type": "string",
            "description": "评分时间",
            "example": "2023-11-20 14:30:00"
          }
        },
        "description": "用户评分实体"
      },
      "ResultRating": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/Rating"
          }
        },
        "description": "统一 API 响应结构"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}